refactor: extract modules by their roles
* store: remains as pure store
* selector: provides ways to select states
* slices: the collection of states & dispatchers for calendar app
* contexts: as a mixed module using Context API. knows how to create context,
            and exports methods to use it.
