+++ apps/calendar/src/components/events/gridEvent.tsx
@@ -5,8 +5,8 @@ import GridWithMouse from '@src/components/daygrid/gridWithMouse';
 import ResizeIcon from '@src/components/events/resizeIcon';
 import { useDrag } from '@src/components/hooks/drag';
 import Template from '@src/components/template';
+import { useDispatch } from '@src/contexts/calendarStore';
 import ScheduleViewModel from '@src/model/scheduleViewModel';
-import { useDispatch } from '@src/store';
 import { cls } from '@src/util/cssHelper';
 import { getGridDateIndex } from '@src/util/gridHelper';
 import { toPercent, toPx } from '@src/util/units';
