+++ apps/calendar/src/components/view/month.tsx
@@ -7,8 +7,9 @@ import { usePanelContainer } from '@src/components/hooks/panelContainer';
 import { useTheme } from '@src/components/hooks/theme';
 import Panel from '@src/components/panel';
 import { MONTH_DAY_NAME_HEIGHT } from '@src/constants/style';
+import { useStore } from '@src/contexts/calendarStore';
 import { MonthOption, TemplateMonthDayName } from '@src/model';
-import { topLevelStateSelector, useStore } from '@src/store';
+import { topLevelStateSelector } from '@src/selectors';
 import { getGridInfo, getMonthCalendar, isWeekend } from '@src/time/datetime';
 import { cls } from '@src/util/cssHelper';
 import { capitalizeDayName } from '@src/util/dayName';
