+++ apps/calendar/src/slices/template.ts
