+++ apps/calendar/test/app/template/template.spec.tsx
@@ -2,9 +2,9 @@ import { h } from 'preact';
 import renderToString from 'preact-render-to-string';
 
 import Template from '@src/components/template';
+import { initCalendarStore, StoreProvider } from '@src/contexts/calendarStore';
 import Schedule from '@src/model/schedule';
 import ScheduleViewModel from '@src/model/scheduleViewModel';
-import { initializeStore, StoreProvider } from '@src/store';
 import {
   getCommonWidth,
   getMonthScheduleBlock,
@@ -32,7 +32,7 @@ describe('Render Template', () => {
   });
 
   it('Template component renders html string with given template', () => {
-    const store = initializeStore();
+    const store = initCalendarStore();
     const vdom = (
       <StoreProvider store={store}>
         <Template template="time" model={{ title: 'Custom Title 4' }} />
